Added code-ranges so the extension does not break if there are more than 2000 tokens in the current file.
NOTE: currently there is an issue where if the cursor is at the top or
the bottom of the editor and there aren't enough tokens above it, it
will simply provide the server with fewer tokens. In the future it
should compensate this by taking more tokens from the other side.