Re-wrote basically everything to make it cleaner and easier to work with.
When I started writing this extension I had no idea how to write
extensions or even typescript. Now that I am much more familiar with
both, I thought it would be worth re-writing it to make sure everything
is cleaner.

I am still a beginner, and its entirely possible that this will need to
be done a couple more times in the future.