Added a toggle at the bottom status bar to turn on and off the authenguard extension
I feel like we should have a more "in your face" toggle to remind people
to use it a bit more. However, im not sure what is best as I also really
don't want it to be annoying.

One idea could be to steal from snyk and add a sidebar item (like where
the extensions and the file manager is)